@@ -31,53 +37,60 @@
     ]
   },
   "dependencies": {
-    "@remix-run/css-bundle": "^1.19.3",
-    "@remix-run/node": "^1.19.3",
-    "@remix-run/react": "^1.19.3",
-    "@remix-run/serve": "^1.19.3",
-    "clsx": "^1.2.1",
-    "isbot": "latest",
+    "@remix-run/css-bundle": "^2.5.1",
+    "@remix-run/node": "^2.5.1",
+    "@remix-run/react": "^2.5.1",
+    "@remix-run/serve": "^2.5.1",
+    "clsx": "^2.1.0",
+    "isbot": "^4.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
-    "@playwright/test": "^1.34.3",
-    "@remix-run/dev": "^1.19.3",
-    "@remix-run/eslint-config": "^1.19.3",
+    "@playwright/test": "^1.41.1",
+    "@remix-run/dev": "^2.5.1",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.9",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/eslint": "^8.37.0",
-    "@types/node": "^18.15.11",
-    "@types/react": "^18.0.37",
-    "@types/react-dom": "^18.0.11",
-    "@vitejs/plugin-react": "^3.1.0",
-    "@vitest/coverage-c8": "^0.30.1",
-    "autoprefixer": "^10.4.14",
+    "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.10",
+    "@testing-library/jest-dom": "^6.4.1",
+    "@testing-library/react": "^14.2.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/eslint": "^8.56.2",
+    "@types/node": "^20.11.15",
+    "@types/react": "^18.2.51",
+    "@types/react-dom": "^18.2.18",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "@vitest/coverage-c8": "^0.33.0",
+    "autoprefixer": "^10.4.17",
     "binode": "^1.0.5",
-    "c8": "^7.13.0",
+    "c8": "^9.1.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.38.0",
-    "eslint-config-prettier": "^8.8.0",
-    "happy-dom": "^9.8.0",
-    "msw": "^1.2.1",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest-dom": "^5.1.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^6.2.0",
+    "happy-dom": "^13.3.8",
+    "msw": "^2.1.5",
     "nano-staged": "^0.8.0",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.22",
-    "prettier": "2.8.7",
-    "prettier-plugin-tailwindcss": "^0.2.7",
+    "postcss": "^8.4.33",
+    "prettier": "^3.2.4",
+    "prettier-plugin-tailwindcss": "^0.5.11",
     "simple-git-hooks": "^2.9.0",
-    "tailwindcss": "^3.3.1",
-    "typescript": "^5.0.4",
-    "vite": "^4.2.1",
-    "vite-tsconfig-paths": "^3.6.0",
-    "vitest": "^0.30.1"
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3",
+    "vite-tsconfig-paths": "^4.3.1",
+    "vitest": "^1.2.2"
   },
   "engines": {
-    "node": ">=14"
+    "node": "^18 || >=20"
   }
 }
