@@ -1,4 +1,5 @@
-module.exports = {
+/** @type {import('postcss-load-config').Config} */
+export default {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
