@@ -1,16 +1,5 @@
 /** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
+export default {
   cacheDirectory: "./node_modules/.cache/remix",
-  future: {
-    v2_errorBoundary: true,
-    v2_meta: true,
-    v2_headers: true,
-    v2_normalizeFormMethod: true,
-    v2_routeConvention: true,
-    v2_dev: true,
-  },
   ignoredRouteFiles: ["**/.*", "**/*.test.{js,jsx,ts,tsx}"],
-  postcss: true,
-  serverModuleFormat: "cjs",
-  tailwind: true,
 };
