@@ -13,8 +13,12 @@ const AnalyticsCommandsContext = createContext<{
   trackPageview(url: string): void;
   trackGoal(eventCode: string, value: number): void;
 }>({
-  trackPageview() {},
-  trackGoal() {},
+  trackPageview() {
+    // do nothing
+  },
+  trackGoal() {
+    // do nothing
+  },
 });
 
 export function useAnalytics() {
@@ -78,13 +82,14 @@ export function AnalyticsScript() {
               queuedCommand.value
             );
             break;
-          default:
+          default: {
             const _: never = queuedCommand;
             console.error(
               `Unknown queued command: ${JSON.stringify(
                 (_ as QueuedCommand).type
               )}`
             );
+          }
         }
       } else {
         // Fathom isn't available even though the script has loaded
