@@ -1,23 +1,4 @@
-import { useMatches } from "@remix-run/react";
-import { useMemo } from "react";
-import type { V2_MetaDescriptor } from "@remix-run/node";
-
-/**
- * This base hook is used in other hooks to quickly search for specific data
- * across all loader data using useMatches.
- * @param {string} id The route id
- * @returns {JSON|undefined} The router data or undefined if not found
- */
-export function useMatchesData(
-  id: string
-): Record<string, unknown> | undefined {
-  const matchingRoutes = useMatches();
-  const route = useMemo(
-    () => matchingRoutes.find((route) => route.id === id),
-    [matchingRoutes, id]
-  );
-  return route?.data;
-}
+import type { MetaDescriptor } from "@remix-run/node";
 
 export function validateEmail(email: unknown): email is string {
   return typeof email === "string" && email.length > 3 && email.includes("@");
@@ -62,7 +43,7 @@ export function getMeta({
   title,
   description,
   image,
-}: PageMeta): V2_MetaDescriptor[] {
+}: PageMeta): MetaDescriptor[] {
   if (description && process.env.NODE_ENV === "development") {
     if (description.length < 110) {
       throw new Error(
