@@ -1,4 +1,4 @@
 import { installGlobals } from "@remix-run/node";
-import "@testing-library/jest-dom";
+import "@testing-library/jest-dom/vitest";
 
 installGlobals();
